Improved error handling code, added check for non-existing routes
An abstract custom error class is added to reduce redundant code and
provide type and format checks to sent error response.
A route handler is added to throw NotFoundError if a non-defined route
is accessed. Express-async-errors pkg support is used to correct the
infinite response time of the server if async calls are used in the
code.