Added a request-validation middleware, made same sanity changes to the code
The Request-validation middleware will reduce duplicate code in every POST/PUT request router logic.
Modified the User-model by adding a JSON-transform function  to hide sensitive fields like hashed password and version when
sending a response to the user during signin/signup process.
Stored the jwt-signing key in a safe location- in the cluster itself
instead of hardcoding it in code, the pods will be able to fetch it as
env-variable when required.