Create tickets db model, route-handler and add lots of tests
The new-ticket schema is created with 3 fields(title,price,userid).
The add new ticket route handler is setup with proper authentication and
validations. A bunch of tests are added to the new route, the global
test setup.ts file is modified to fake user-sign in by generating valid cookies using the jwt signing key.